+++ src/app/aura.cpp
@@ -13,6 +13,7 @@ using namespace Nickvision::Filesystem;
 using namespace Nickvision::Helpers;
 using namespace Nickvision::Localization;
 using namespace Nickvision::Logging;
+using namespace Nickvision::Notifications;
 using namespace Nickvision::System;
 
 namespace Nickvision::App
@@ -214,6 +215,17 @@ namespace Nickvision::App
         }
         return "https://htmlpreview.github.io/?" + m_appInfo.getHtmlDocsStore() + "/" + lang + "/" + pageName + ".html";
     }
+
+#ifdef _WIN32
+    NotifyIcon& Aura::getNotifyIcon(HWND hwnd)
+    {
+        if (!m_notifyIcon)
+        {
+            m_notifyIcon = std::make_unique<NotifyIcon>(hwnd, StringHelpers::wstr(m_appInfo.getName()), NotifyIconMenu(), true);
+        }
+        return *m_notifyIcon;
+    }
+#endif
 
     Aura::operator bool() const
     {